Rename DelayedGameEvent.cs to DelayedLevelFailed.cs
